requirements.txt
django==5.1
djangorestframework==3.15.2
openpyxl==3.1.5
pandas==2.2.3
django-cors-headers==4.6.0
catboost==1.2.7
scikit-learn==1.6.0